Make 'linear' module a package, complete 'Contour.__contains__' magic method
