+++ gon/linear/__init__.py
@@ -1,0 +1,5 @@
+from .contour import Contour
+from .hints import (RawContour,
+                    RawSegment,
+                    Vertices)
+from .segment import Segment
