+++ gon/linear/hints.py
@@ -1,0 +1,10 @@
+from typing import (List,
+                    Sequence,
+                    Tuple)
+
+from gon.primitive import (Point,
+                           RawPoint)
+
+RawSegment = Tuple[RawPoint, RawPoint]
+RawContour = List[RawPoint]
+Vertices = Sequence[Point]
